% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_citation.R
\name{format_apa}
\alias{format_apa}
\title{Format citation in APA style}
\usage{
format_apa(bibtex_data, link_doi = TRUE)
}
\arguments{
\item{bibtex_data}{A list containing BibTeX fields for a publication.}

\item{link_doi}{Logical indicating whether DOIs should be formatted as
hyperlinks. Default is TRUE.}
}
\value{
A character string containing the APA-formatted citation, or NA if
formatting fails.
}
\description{
Creates a citation formatted according to American Psychological Association
(APA) style guidelines from BibTeX data.
}
\details{
APA style formatting includes:
\itemize{
\item Authors: last name, initials format with "&" before final author
\item For 1-7 authors: all authors listed
\item For 8+ authors: first 6 authors, ellipsis, last author
\item Year in parentheses after authors
\item Title in sentence case without quotes
\item Journal name and volume in italics
\item Issue in parentheses (not italicized)
}
}
\examples{
# create BibTeX data
bibtex_data <- list(
  entry_type = "article",
  author = "Smith, J. and Johnson, S.",
  year = "2023",
  title = "Example research paper",
  journal = "Journal of Examples",
  volume = "45",
  number = "2",
  pages = "123-145",
  doi = "10.1234/example"
)

format_apa(bibtex_data)

}
\seealso{
\code{\link{format_citation}} for the main citation formatting function.
}
