% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_citation.R
\name{format_chicago}
\alias{format_chicago}
\title{Format citation in Chicago style}
\usage{
format_chicago(bibtex_data, link_doi = TRUE)
}
\arguments{
\item{bibtex_data}{A list containing BibTeX fields for a publication.}

\item{link_doi}{Logical indicating whether DOIs should be formatted as
hyperlinks. Default is TRUE.}
}
\value{
A character string containing the Chicago-formatted citation, or NA
if formatting fails.
}
\description{
Creates a citation formatted according to Chicago Manual of Style guidelines
from BibTeX data.
}
\details{
Chicago style formatting includes:
\itemize{
\item Authors: first author in "Last, First" format, subsequent authors in "First Last" format
\item For 1-3 authors: all authors listed
\item For 4+ authors: first author et al.
\item Year after authors
\item Title in quotation marks
\item Journal name in italics
\item Format: Author(s). Year. "Title." Journal Volume, no. Issue: Pages.
}
}
\examples{
# create BibTeX data
bibtex_data <- list(
  entry_type = "article",
  author = "Smith, John and Johnson, Sarah",
  year = "2023",
  title = "Example research paper",
  journal = "Journal of Examples",
  volume = "45",
  number = "2",
  pages = "123-145",
  doi = "10.1234/example"
)

format_chicago(bibtex_data)

}
\seealso{
\code{\link{format_citation}} for the main citation formatting function.
}
