% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_citation.R
\name{format_nature_authors}
\alias{format_nature_authors}
\title{Format author string in Nature style}
\usage{
format_nature_authors(author_str)
}
\arguments{
\item{author_str}{A character string containing author names in "Last, First"
format, separated by " and ".}
}
\value{
A character string containing the authors formatted in Nature style.
}
\description{
Formats author names according to Nature journal style guidelines.
}
\details{
Nature author formatting rules:
\itemize{
\item Single author: "Last, F." with period
\item 2-5 authors: "Last, F., Last, F. & Last, F."
\item 6+ authors: "Last, F. et al."
}
}
\examples{
\dontrun{
format_nature_authors("Smith, John")
format_nature_authors("Smith, John and Johnson, Sarah and Lee, David")
}

}
\seealso{
\code{\link{format_nature}} which uses this function to format citations.
}
\keyword{internal}
