% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_citation.R
\name{format_science_authors}
\alias{format_science_authors}
\title{Format author string in Science style}
\usage{
format_science_authors(author_str)
}
\arguments{
\item{author_str}{A character string containing author names in "Last, First"
format, separated by " and ".}
}
\value{
A character string containing the authors formatted in Science style.
}
\description{
Formats author names according to Science journal style guidelines.
}
\details{
Science author formatting rules:
\itemize{
\item Single author: last name only
\item Multiple authors: first author's last name followed by "et al"
}
}
\examples{
\dontrun{
format_science_authors("Smith, John")
format_science_authors("Smith, John and Johnson, Sarah and Lee, David")
}

}
\seealso{
\code{\link{format_science}} which uses this function to format citations.
}
\keyword{internal}
